// frontend/src/api_gen/models/expense-report-info-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DailyWageDTO } from './daily-wage-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { ExpenseCategoryDTO } from './expense-category-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { TripDTO } from './trip-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { UserDetailsDTO } from './user-details-dto';

/**
 * 
 * @export
 * @interface ExpenseReportInfoDTO
 */
export interface ExpenseReportInfoDTO {
    /**
     * 
     * @type {UserDetailsDTO}
     * @memberof ExpenseReportInfoDTO
     */
    'user': UserDetailsDTO;
    /**
     * 
     * @type {TripDTO}
     * @memberof ExpenseReportInfoDTO
     */
    'trip': TripDTO;
    /**
     * 
     * @type {Array<ExpenseCategoryDTO>}
     * @memberof ExpenseReportInfoDTO
     */
    'expenseCategory': Array<ExpenseCategoryDTO>;
    /**
     * 
     * @type {DailyWageDTO}
     * @memberof ExpenseReportInfoDTO
     */
    'dailyWage': DailyWageDTO;
    /**
     * 
     * @type {number}
     * @memberof ExpenseReportInfoDTO
     */
    'eurExpenseSum': number;
}